requirements.txt
apache-airflow==2.7.2
pandas
psycopg2-binary
requests